feat: add styles and refactor
