+++ client/src/app/page.tsx
@@ -1,1 +1,1 @@
-export { HomePage as default, metadata } from "@presentation/pages/HomePage";
+export { HomePage as default, metadata } from '@presentation/pages/HomePage';
