+++ client/src/presentation/assets/styles/_base.scss
@@ -1,0 +1,11 @@
+body {
+  padding: 0;
+  margin: 0;
+  height: 100vh;
+}
+
+a:link,
+a:visited {
+  text-decoration: none;
+  color: white;
+}
