+++ client/src/presentation/assets/styles/main.scss
@@ -1,1 +1,2 @@
+@import './base';
 @import './partial-imports';
