+++ client/src/presentation/components/Header/header.scss
@@ -1,5 +1,7 @@
 .header {
-    grid-area: header;
-    background-color: red;
-    color: white;
+  grid-area: header;
+  background-color: #121212;
+  color: white;
+  height: 80px;
+  text-align: center;
 }
