+++ client/src/presentation/components/Loader/Loader.tsx
@@ -1,0 +1,14 @@
+import { FC } from 'react';
+
+export const Loader: FC = () => {
+  return (
+    <div className="loader">
+      <div className="loader__bar"></div>
+      <div className="loader__bar"></div>
+      <div className="loader__bar"></div>
+      <div className="loader__bar"></div>
+      <div className="loader__bar"></div>
+      <div className="loader__ball"></div>
+    </div>
+  );
+};
