+++ client/src/presentation/components/MainView/main-view.scss
@@ -1,5 +1,11 @@
 .main-view {
-    grid-area: main-view;
-    background-color: white;
-    color: red;
+  grid-area: main-view;
+  background-color: #121212;
+  color: white;
+  min-height: calc(((100vh - 64px) - 90px) - 92px);
+  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
