+++ client/src/presentation/components/PlayingBar/playing-bar.scss
@@ -1,5 +1,7 @@
 .playing-bar {
-    grid-area: playing-bar;
-    background-color: green;
-    color: white;
+  grid-area: playing-bar;
+  background-color: #121212;
+  color: white;
+  height: 150px;
+  text-align: center;
 }
