+++ client/src/presentation/components/SideBar/SideBar.tsx
@@ -1,9 +1,14 @@
-import { FC } from "react"
+import { FC } from 'react';
+import Link from 'next/link';
 
 export const SideBar: FC = () => {
-    return (
-      <aside className="sidebar">
-        SideBar
-      </aside>
-    )
-}
+  return (
+    <aside className="sidebar">
+      <Link className="link" href="/">
+        Home
+      </Link>
+      <br />
+      SideBar
+    </aside>
+  );
+};
