+++ client/src/presentation/components/SideBar/sidebar.scss
@@ -1,5 +1,7 @@
 .sidebar {
-    grid-area: sidebar;
-    background-color: blue;
-    color: white;
+  grid-area: sidebar;
+  background-color: #121212;
+  color: white;
+  width: 400px;
+  text-align: center;
 }
