+++ client/src/presentation/layouts/RootLayout/RootLayout.tsx
@@ -1,17 +1,17 @@
 import { FC, PropsWithChildren } from 'react';
 
-import { Header } from '@presentation/components/Header'
-import { SideBar } from '@presentation/components/SideBar'
-import { PlayingBar } from '@presentation/components/PlayingBar'
+import { Header } from '@presentation/components/Header';
+import { SideBar } from '@presentation/components/SideBar';
+import { PlayingBar } from '@presentation/components/PlayingBar';
 import { MainView } from '@presentation/components/MainView';
 
-import '../../assets/styles/main.scss'
+import '../../assets/styles/main.scss';
 
 export const RootLayout: FC<PropsWithChildren> = ({ children }) => {
   return (
     <html lang="en">
       <body>
-        <div className='root-content'>
+        <div className="root-content">
           <Header />
           <SideBar />
           <PlayingBar />
