--- client/src/presentation/pages/RootLayout/root-layout.scss
@@ -1,11 +1,0 @@
-.root-content {
-    display: grid;
-    grid-gap: 8px;
-    grid-template-columns: auto 1fr;
-    grid-template-rows: 1fr auto;
-    grid-template-areas: "sidebar header header" "sidebar main-view main-view" "playing-bar playing-bar playing-bar";
-    height: 100%;
-    min-height: 100%;
-    width: 100%;
-    position: relative;
-}
