--- client/src/presentation/pages/TrackPage/track-page.scss
@@ -1,12 +1,0 @@
-.content {
-    display: flex;
-    flex-direction: row;
-
-    & > aside {
-        width: 30%;
-    }
-
-    & > main {
-        width: 70%;
-    }
-}
